Refactor QStringList to qtcore::qlist::qstringlist
