+++ qttypes/src/qtcore/mod.rs
@@ -1,11 +1,13 @@
 mod primitives;
 mod qbytearray;
 mod qchar;
+mod qlist;
 mod qstring;
 mod qurl;
 
 pub use self::primitives::qreal;
 pub use self::qbytearray::QByteArray;
 pub use self::qchar::UnicodeVersion;
+pub use self::qlist::QStringList;
 pub use self::qstring::{NormalizationForm, QString};
 pub use self::qurl::QUrl;
