+++ qttypes/src/qtcore/qlist/mod.rs
@@ -1,0 +1,3 @@
+mod qstringlist;
+
+pub use self::qstringlist::QStringList;
